+++ NEWS.md
@@ -9,6 +9,7 @@ User-visible changes worth mentioning.
 - [#970] Escape certain attributes in authorization forms.
 - [#974] Redirect URI is checked without query params within AuthorizationCodeRequest.
 - [#1023] Update Ruby versions and test against 2.5.0 on Travis CI.
+- [#1024] Migrate from FactoryGirl to FactoryBot.
 
 ## 4.2.5